requirements.txt
anthropic==0.42.0
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
streamlit
numpy==1.26.4